[WIP] factorize occupancy functions
